refactor: use forward declerations more than includes
