+++ include/command.h
@@ -1,7 +1,9 @@
 #ifndef COMMAND_H
 #define COMMAND_H
 
-#include "container.h"
+#include <wlr/types/wlr_seat.h>
+
+struct container;
 
 typedef struct cmd_results *sway_cmd(int argc, char **argv);
 
@@ -49,7 +51,9 @@ struct cmd_results *cmd_results_new(enum cmd_status status, const char *error, .
  */
 void free_cmd_results(struct cmd_results *results);
 
-struct cmd_results *execute_command(char *cmd, struct wlr_seat *seat,
+struct cmd_results *execute_command(
+        char *cmd,
+        struct wlr_seat *seat,
         struct container *con);
 char *cmd_results_to_json(struct cmd_results *results);
